Refactoring triple headerr diretory structure
