+++ packages/triple-header/src/frame/effects/common.ts
