+++ packages/triple-header/src/frame/effects/none.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react'
 
-import { MotionContainer } from '../motion-container'
+import { MotionContainer } from '../../motion-container'
 
 export default function None({ children }: PropsWithChildren) {
   return <MotionContainer>{children}</MotionContainer>
