+++ packages/triple-header/src/frame/index.ts
@@ -1,7 +1,10 @@
 import ImageFrame from './image'
 import TextFrame from './text'
+import Frame from './frame'
 
 export const FRAMES = {
   image: ImageFrame,
   text: TextFrame,
 }
+
+export default Frame
