+++ packages/triple-header/src/layer/index.ts
@@ -1,0 +1,3 @@
+import Layers from './layer'
+
+export default Layers
