+++ packages/triple-header/src/layer/transitions/marquee.tsx
