+++ packages/triple-header/src/triple-header.tsx
@@ -1,10 +1,9 @@
-import { useState, ComponentType, useCallback, useLayoutEffect } from 'react'
+import { useState, useCallback, useLayoutEffect } from 'react'
 import { Container } from '@titicaca/core-elements'
 import styled, { css } from 'styled-components'
 
-import { TRANSITIONS } from './transitions'
-import { FRAMES } from './frames'
-import { FrameData, TripleHeader as TripleHeaderProps } from './types'
+import Layer from './layer'
+import { TripleHeader as TripleHeaderProps } from './types'
 
 const Canvas = styled(Container).attrs({
   position: 'relative',
@@ -20,42 +19,6 @@ const Canvas = styled(Container).attrs({
       width: 100%;
       height: calc(${clientWidth}px * ${height / width});
       max-height: ${768 * (height / width)}px;
-    `}
-`
-
-const Layer = styled(Container).attrs({
-  position: 'absolute',
-})<{
-  zIndex: number
-}>`
-  width: 100%;
-
-  ${({ zIndex }) =>
-    zIndex &&
-    css`
-      z-index: ${zIndex};
-    `}
-`
-
-const FrameContainer = styled(Container)<{
-  widthRatio: number
-  heightRatio: number
-}>`
-  width: 100%;
-  height: 0;
-  margin: 0 auto;
-
-  ${({ heightRatio }) =>
-    heightRatio &&
-    css`
-      padding: ${heightRatio}% 0 0 0;
-      position: relative;
-    `}
-
-  ${({ widthRatio }) =>
-    widthRatio &&
-    css`
-      max-width: ${widthRatio}%;
     `}
 `
 
@@ -90,6 +53,10 @@ export default function TripleHeader({
 
   const { canvas, layers } = children
 
+  const calculateFrameRatio = (length?: number) => {
+    return length ? (length / canvas.width) * 100 : 0
+  }
+
   return canvas && layers ? (
     <Canvas
       ref={previewRef}
@@ -98,10 +65,6 @@ export default function TripleHeader({
       height={canvas.height}
     >
       {layers.map(({ frames, transition, positioning }, index) => {
-        const LayerElement = transition
-          ? TRANSITIONS[transition.type]
-          : Container
-
         const position = {
           top: (Number(positioning?.top || 0) / canvas.height) * 100,
           left: (Number(positioning?.left || 0) / canvas.height) * 100,
@@ -111,32 +74,11 @@ export default function TripleHeader({
           <Layer
             key={index}
             zIndex={index + 1}
-            css={{
-              top: `${position.top}%`,
-              left: `${position.left}%`,
-            }}
-          >
-            <LayerElement>
-              {frames.map(({ type, width, height, value, effect }, index) => {
-                const FrameElement = FRAMES[type] as ComponentType<
-                  Omit<FrameData, 'type'>
-                >
-
-                const widthRatio = width ? (width / canvas.width) * 100 : 0
-                const heightRatio = height ? (height / canvas.width) * 100 : 0
-
-                return (
-                  <FrameContainer
-                    key={index}
-                    widthRatio={widthRatio}
-                    heightRatio={heightRatio}
-                  >
-                    <FrameElement value={value} effect={effect} />
-                  </FrameContainer>
-                )
-              })}
-            </LayerElement>
-          </Layer>
+            position={position}
+            frames={frames}
+            transition={transition}
+            calculateFrameRatio={calculateFrameRatio}
+          />
         )
       })}
     </Canvas>
